Now using BigIntegerField to store schedule hash.

+++ periodically/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 class ExecutionRecord(models.Model):
     task_id = models.CharField(max_length=255)
-    schedule_id = models.IntegerField()
+    schedule_id = models.BigIntegerField()
     scheduled_time = models.DateTimeField()
     start_time = models.DateTimeField()
     end_time = models.DateTimeField(blank=True, null=True)
